adding middlewares

+++ src/Controllers/todoController.ts
@@ -43,21 +43,36 @@ export const create = async (req: Request, res: Response) => {
 export const updateT = async (req: Request, res: Response) => {
   try {
     const TodoId = req.params.id;
-
-    const updatedTodo = await Todo.findByIdAndUpdate(
-      { _id: TodoId },
-      {
-        title: req.body.title,
-        description: req.body.description,
-        completed: req.body.completed,
-      },
-      { new: true }
-    );
-
-    return res.status(200).json({
-      updatedTodo,
-      message: "your todo was successfully updated",
-    });
+    const todo = await Todo.findById(TodoId);
+    if (!req.user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+    if (!todo) {
+      return res.status(400).json({
+        status: "failed",
+        message: "Id of a todo not found",
+      });
+    }
+    if (todo?.addedBy === req.user.email) {
+      const updatedTodo = await Todo.findByIdAndUpdate(
+        { _id: TodoId },
+        {
+          title: req.body.title,
+          description: req.body.description,
+          completed: req.body.completed,
+        },
+        { new: true }
+      );
+      return res.status(200).json({
+        updatedTodo,
+        message: "your todo was successfully updated",
+      });
+    } else {
+      return res.status(400).json({
+        status: "failed",
+        message: "It is not belonged to you",
+      });
+    }
   } catch (error) {
     return res.status(500).json(error);
   }
@@ -67,17 +82,28 @@ export const updateT = async (req: Request, res: Response) => {
 export const deleteT = async (req: Request, res: Response) => {
   try {
     const id = req.params.id;
-    const todo = await Todo.findByIdAndDelete(id);
+    if (!req.user) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+    const todo = await Todo.findById(id);
     if (!todo) {
       return res.status(400).json({
         status: "failed",
         message: "Id of a todo not found",
       });
     }
-    return res.status(204).json({
-      status: "success",
-      message: "Todo deleted ............",
-    });
+    if (todo?.addedBy === req.user.email) {
+      await Todo.findByIdAndDelete(id);
+      return res.status(204).json({
+        status: "success",
+        message: "Todo deleted ............",
+      });
+    } else {
+      return res.status(400).json({
+        status: "failed",
+        message: "It is not bellonged to you",
+      });
+    }
   } catch (err) {
     return res.status(500).json(err);
   }
